General Upload of what I did
Style contains all the jQueryUI images and style layouts. I added a
little style in the file main.css all the magic happens in lib/color.js!
All the data from the sliders get converted into a css line, which will
add the color to the background of your page.